use new scheduler
