+++ .gitignore
@@ -15,3 +15,4 @@ _*
 
 !__init__.py
 !__main__.py
+/.*rules
