+++ src/gary/llm/__init__.py
@@ -1,6 +1,6 @@
-from .llm import LLM
-from .scheduler import Scheduler
+from .llm import LLM, Act
+from .scheduler2 import Scheduler2
 from .randy import Randy
 from .llarry import Llarry
 
-__all__ = ["LLM", "Scheduler", "Llarry", "Randy"]
+__all__ = ["LLM", "Act", "Scheduler2", "Llarry", "Randy"]
