+++ src/gary/util/__init__.py
@@ -2,6 +2,7 @@ from .config import CONFIG
 from .logger import logger
 from .websocket import WSConnection, GameWSConnection
 from .utils import invoke, HasEvents, NoPrint
+from .periodic_timer import PeriodicTimer
 
 __all__ = [
     "CONFIG",
@@ -11,4 +12,5 @@ __all__ = [
     "invoke",
     "HasEvents",
     "NoPrint",
+    "PeriodicTimer",
 ]
